+++ userprofile/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class UserprofileConfig(AppConfig):
-    name = 'user_profile'
+    name = 'userprofile'