+++ userprofile/plugins.py
@@ -8,7 +8,7 @@ class SiteAppPlugin(SiteAppPluginPoint):
     """Projectroles plugin for registering the app"""
 
     #: Name (slug-safe, used in URLs)
-    name = 'user_profile'
+    name = 'userprofile'
 
     #: Title (used in templates)
     title = 'User Profile'
@@ -23,10 +23,10 @@ class SiteAppPlugin(SiteAppPluginPoint):
     description = 'Project User Profile'
 
     #: Entry point URL ID
-    entry_point_url_id = 'user_profile:detail'
+    entry_point_url_id = 'userprofile:detail'
 
     #: Required permission for displaying the app
-    app_permission = 'user_profile.view_detail'
+    app_permission = 'userprofile.view_detail'
 
     def get_messages(self, user=None):
         """